Add -Wconversion flag and fix errors.

+++ connection.h
@@ -57,7 +57,8 @@ struct Connection {
             return false;
         }
         // io_uring_prep_read(sqe, fd, NextBufferToRead(), NextReadLengthLimit(), 0);
-        io_uring_prep_read(sqe, fd, buffer.AppendBuffer().data(), buffer.AppendLen(), 0);
+        io_uring_prep_read(
+          sqe, fd, buffer.AppendBuffer().data(), static_cast<unsigned int>(buffer.AppendLen()), 0);
         io_uring_sqe_set_data(sqe, AsData());
         if (SQE_ASYNC) {
             io_uring_sqe_set_flags(sqe, IOSQE_ASYNC);
@@ -172,7 +173,8 @@ struct Connection {
         if (sqe == nullptr) {
             return false;
         }
-        io_uring_prep_write(sqe, fd, query_buffer.data(), query_buffer.size(), 0);
+        io_uring_prep_write(
+          sqe, fd, query_buffer.data(), static_cast<unsigned int>(query_buffer.size()), 0);
         io_uring_sqe_set_data(sqe, AsData());
         if (SQE_ASYNC) {
             io_uring_sqe_set_flags(sqe, IOSQE_ASYNC);
@@ -209,13 +211,9 @@ struct Connection {
         SetClosing();
     }
 
-    void SetClosing() {
-        state = State::Closing;
-    }
+    void SetClosing() { state = State::Closing; }
 
-    bool Alive() const {
-        return state == State::Alive;
-    }
+    bool Alive() const { return state == State::Alive; }
 
     std::string Command() const { return vec[0].GetString(); }
 };
